clean up deps

+++ indexer/src/storage/lm_db/types.rs
@@ -3,5 +3,5 @@ pub struct LMDBClientConfig {
     pub db_name: String,
     pub path: String,
     pub max_dbs: u32,
-    pub map_size: u32,
+    pub map_size: usize,
 }
